requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
-e git+https://github.com/kyb3r/dhooks.git@20a79f42fe799470fe8d079f877316475dc867bc#egg=dhooks
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.12
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
oauthlib==3.1.0
pbr==5.5.1
Pillow==8.1.0
pycodestyle==2.6.0
PySocks==1.7.1
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
smmap==3.0.4
stevedore==3.3.0
toml==0.10.2
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
